[CIR] Spell the identity as std::find, tighten the recognizer and tests
Print the identity spelling as std::find. It is not a bare identifier,
so a quoting printer serializes it and the enum parser reads it back.

Fold the two func_info setters into one setFuncInfoAttr. The special
member forms come first and return, and a known standard library entity
is the fall through, so a member and a free function never both fire.

Move the operand and result count checks into the per operation shape
check, now signatureMatches, so each check validates its own shape
before it reads the operands.

Tighten the tests. The guards pin the full mangled name, the lowered
call is checked with its operands bound to their named allocas so a
reordering would fail, the declined cases assert both the non raise and
the survival, and the parsed CIR test asserts the operand order. The
namespace cases join the guards file, since they decline for the same
reason.
