+++ clang/lib/CIR/CodeGen/CIRGenModule.h
@@ -798,14 +798,11 @@ public:
                                        cir::FuncType ty,
                                        const clang::FunctionDecl *fd);
 
-  /// Mark the function as a special member (e.g. constructor, destructor)
-  void setCXXSpecialMemberAttr(cir::FuncOp funcOp,
-                               const clang::FunctionDecl *funcDecl);
-
-  /// Tag functions that match a known standard library entity, so passes
-  /// can recognize calls to them without the AST.
-  void setFuncIdentityAttr(cir::FuncOp funcOp,
-                           const clang::FunctionDecl *funcDecl);
+  /// Record the func_info tag for a function, either a C++ special member
+  /// form (constructor, destructor, assignment) or a known standard library
+  /// entity that passes can recognize without the AST.
+  void setFuncInfoAttr(cir::FuncOp funcOp,
+                       const clang::FunctionDecl *funcDecl);
 
   cir::FuncOp createRuntimeFunction(cir::FuncType ty, llvm::StringRef name,
                                     mlir::NamedAttrList extraAttrs = {},
