+++ clang/test/CIR/IR/func-identity-attr.cir
@@ -2,7 +2,7 @@
 
 module {
 
-cir.func private @_ZSt4findv() func_info<#cir.func_identity<std_find>>
-// CHECK: cir.func private @_ZSt4findv() func_info<#cir.func_identity<std_find>>
+cir.func private @_ZSt4findv() func_info<#cir.func_identity<"std::find">>
+// CHECK: cir.func private @_ZSt4findv() func_info<#cir.func_identity<"std::find">>
 
 }
