+++ clang/test/CIR/Transforms/idiom-recognizer-guards.cpp
@@ -36,3 +36,32 @@ char *test_wrong_arg_count(char *first, char *last, const char &value) {
 }
 // CHECK-LABEL: @_Z20test_wrong_arg_count
 // CHECK: cir.call @_ZSt4findPcS_RKci
+
+// std membership is fixed when the tag is set, so a find outside std is never
+// tagged and never raised, whatever the call shape.
+
+// A nested namespace that is not inline is not std.
+namespace std {
+namespace another_ns {
+template <class Iter, class T>
+Iter find(Iter, Iter, const T &);
+}
+}
+
+char *test_nested_namespace(char *first, char *last, const char &value) {
+  return std::another_ns::find(first, last, value);
+}
+// CHECK-LABEL: @_Z21test_nested_namespace
+// CHECK: cir.call
+
+// An anonymous namespace function is not std::find.
+namespace {
+template <class Iter, class T>
+Iter find(Iter, Iter, const T &);
+}
+
+char *test_anonymous_namespace(char *first, char *last, const char &value) {
+  return find(first, last, value);
+}
+// CHECK-LABEL: @_Z24test_anonymous_namespace
+// CHECK: cir.call
